requirements.txt
# main dependencies
langchain
langchain-openai
langchain_community

# extra dependencies
python-dotenv

# test dependencies
pytest

# format dependencies
black
